+++ libs/zhi-fetch-middleware/src/lib/commonFetchClient.ts
@@ -113,7 +113,7 @@ class CommonFetchClient {
     this.logger.info(`check response is stream =>${isStream}`)
 
     if (isStream) {
-      this.logger.info("检测到response不是Response的实例，直接返回", typeof response)
+      this.logger.info("检测到response不是Response的实例", typeof response)
       resJson = response
     } else {
       // 解析响应体并返回响应结果