+++ libs/zhi-blog-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # zhi-blog-api
 
+## 1.56.3
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-common@1.32.0
+
 ## 1.56.2
 
 ### Patch Changes