+++ libs/zhi-gitlab-middleware/CHANGELOG.md
@@ -1,5 +1,31 @@
 # zhi-gitlab-middleware
 
+## 0.7.0
+
+### Minor Changes
+
+- feat: add asset api
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-fetch-middleware@0.9.0
+
+## 0.6.28
+
+### Patch Changes
+
+- fix: revert resp parse
+- Updated dependencies
+  - zhi-fetch-middleware@0.8.2
+
+## 0.6.27
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-fetch-middleware@0.8.1
+
 ## 0.6.26
 
 ### Patch Changes